BaseInventory: Move contents of InventoryHelpersTrait back inline I moved these to a trait in anticipation of having multiple full Inventory implementations. That's no longer necessary because of the abstraction of SimpleInventory.
